task2- Updated the values of number fields in db and services to be ints instead of doubles. This now allows for appropriate output and storage in the database, including BigInteger for the totalLCoE field, where the result could be billions.

+++ ostracon-project/src/main/java/ostracon/ostracon_project/power_plants/CalculationsService.java
@@ -1,5 +1,7 @@
 package ostracon.ostracon_project.power_plants;
 
+import java.math.BigInteger;
+
 /**
  * Services for calculating a power plants electricity generated annually,
  * direct emissions of carbon dioxide, global warming potential and total LCoE.
@@ -14,7 +16,7 @@ public interface CalculationsService
 	 * @param powerPlant
 	 * @return electricity generated annually
 	 */
-	public double electricityGeneratedAnnually(PowerPlant powerPlant);
+	public int electricityGeneratedAnnually(PowerPlant powerPlant);
 	
 	/**
 	 * To calculate the direct emissions of CO2, based on a power plants 
@@ -22,7 +24,7 @@ public interface CalculationsService
 	 * @param powerPlant
 	 * @return direct emissions of CO2
 	 */
-	public double directEmissionsOfCarbonDioxide(PowerPlant powerPlant);
+	public int directEmissionsOfCarbonDioxide(PowerPlant powerPlant);
 	
 	/**
 	 * To calculate the global warming potential of a power plant, based on its 
@@ -30,7 +32,7 @@ public interface CalculationsService
 	 * @param powerPlant
 	 * @return global warming potential
 	 */
-	public double globalWarmingPotential(PowerPlant powerPlant);
+	public int globalWarmingPotential(PowerPlant powerPlant);
 	
 	/**
 	 * To calculate the total LCoE of a power plant, based on its
@@ -38,7 +40,7 @@ public interface CalculationsService
 	 * @param powerPlant
 	 * @return total LCoE
 	 */
-	public double totalLcoe(PowerPlant powerPlant);
+	public BigInteger totalLcoe(PowerPlant powerPlant);
 	
 	/**
 	 * For returning the total sum of the electricity generated annually
@@ -47,7 +49,7 @@ public interface CalculationsService
 	 * @param year
 	 * @return electricity generated annually
 	 */
-	public double totalElectricityGeneratedAnnuallyByCountryandYear(String country, Integer year);
+	public int totalElectricityGeneratedAnnuallyByCountryandYear(String country, Integer year);
 	
 	/**
 	 * For returning the total sum of direct emissions of CO2
@@ -56,7 +58,7 @@ public interface CalculationsService
 	 * @param year
 	 * @return direct emissions of CO2
 	 */
-	public double totalDirectEmissionsOfCarbonDioxideByCountryAndYear(String country, Integer year);
+	public int totalDirectEmissionsOfCarbonDioxideByCountryAndYear(String country, Integer year);
 	
 	/**
 	 * For returning the total sum of global warming potential
@@ -65,7 +67,7 @@ public interface CalculationsService
 	 * @param year
 	 * @return global warming potential
 	 */
-	public double totalGlobalWarmingPotentialByCountryAndYear(String country, Integer year);
+	public int totalGlobalWarmingPotentialByCountryAndYear(String country, Integer year);
 	
 	/**
 	 * For returning the total sum of all total LCoE's
@@ -74,6 +76,6 @@ public interface CalculationsService
 	 * @param year
 	 * @return total LCoE
 	 */
-	public double totalOfAllLcoeByCountryAndYear(String country, Integer year);
+	public BigInteger totalOfAllLcoeByCountryAndYear(String country, Integer year);
 	
 }
